+++ tiles/src/main/java/org/apache/struts/tiles/TilesRequestProcessor.java
@@ -27,12 +27,12 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 import org.apache.struts.action.ActionServlet;
 import org.apache.struts.action.RequestProcessor;
 import org.apache.struts.config.ForwardConfig;
 import org.apache.struts.config.ModuleConfig;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * <p><strong>RequestProcessor</strong> contains the processing logic that
@@ -60,9 +60,9 @@ public class TilesRequestProcessor extends RequestProcessor {
     protected DefinitionsFactory definitionsFactory = null;
 
     /**
-     * Commons Logging instance.
+     * SLF4J Logging instance.
      */
-    protected static Log log = LogFactory.getLog(TilesRequestProcessor.class);
+    protected static Logger log = LoggerFactory.getLogger(TilesRequestProcessor.class);
 
     /**
      * Initialize this request processor instance.
@@ -94,19 +94,15 @@ public class TilesRequestProcessor extends RequestProcessor {
 
         if (definitionsFactory == null) { // problem !
 
-            log.info(
-                "Definition Factory not found for module '"
-                    + moduleConfig.getPrefix()
-                    + "'. "
-                    + "Have you declared the appropriate plugin in struts-config.xml ?");
+            log.info("Definition Factory not found for module '{}'. "
+                + "Have you declared the appropriate plugin in struts-config.xml ?",
+                moduleConfig.getPrefix());
 
             return;
         }
 
-        log.info(
-            "Tiles definition factory found for request processor '"
-                + moduleConfig.getPrefix()
-                + "'.");
+        log.info("Tiles definition factory found for request processor '{}'.",
+            moduleConfig.getPrefix());
 
     }
 
@@ -190,7 +186,7 @@ public class TilesRequestProcessor extends RequestProcessor {
                             getServletContext());
                 } catch (NoSuchDefinitionException ex) {
                     // Ignore not found
-                    log.debug("NoSuchDefinitionException " + ex.getMessage());
+                    log.debug("NoSuchDefinitionException {}", ex.getMessage());
                 }
                 if (definition != null) { // We have a definition.
                     // We use it to complete missing attribute in context.
@@ -263,9 +259,7 @@ public class TilesRequestProcessor extends RequestProcessor {
 
         // If request comes from a previous Tile, do an include.
         // This allows to insert an action in a Tile.
-        if (log.isDebugEnabled()) {
-            log.debug("uri=" + uri + " doInclude=" + doInclude);
-        }
+        log.debug("uri={} doInclude={}" , uri, doInclude);
 
         if (doInclude) {
             doInclude(uri, request, response);
@@ -323,27 +317,16 @@ public class TilesRequestProcessor extends RequestProcessor {
             return;
         }
 
-        if (log.isDebugEnabled()) {
-            log.debug(
-                "processForwardConfig("
-                    + forward.getPath()
-                    + ")");
-        }
+        log.debug("processForwardConfig({})", forward.getPath());
 
         // Try to process the definition.
         if (processTilesDefinition(forward.getPath(),
-            request,
-            response)) {
-            if (log.isDebugEnabled()) {
-                log.debug(
-                    "  '" + forward.getPath() + "' - processed as definition");
-            }
+                request, response)) {
+            log.debug("  '{}' - processed as definition", forward.getPath());
             return;
         }
 
-        if (log.isDebugEnabled()) {
-            log.debug("  '" + forward.getPath() + "' - processed as uri");
-        }
+        log.debug("  '{}' - processed as uri", forward.getPath());
 
         // forward doesn't contain a definition, let parent do processing
         super.processForwardConfig(request, response, forward);
@@ -406,5 +389,4 @@ public class TilesRequestProcessor extends RequestProcessor {
     public DefinitionsFactory getDefinitionsFactory() {
         return definitionsFactory;
     }
-
 }